Prep of full stack design
