+++ exa.py
@@ -4,8 +4,8 @@ exa Launcher
 ===================
 Useful for development and testing
 '''
-#import sys
-#import os
-#sys.path.insert(0, os.path.abspath('./'))
+import sys
+import os
+sys.path.insert(0, os.path.abspath('.'))
 from exa.__main__ import main
 main()
