+++ exa/static/css/dashboard.css
@@ -3,48 +3,3 @@ body {
   font-family: 'Roboto', 'Helvetica Neue', Helvetica, Arial, sans-serif;
   color: #333;
 }
-/*
-header {
-    background-color: #000000;
-    color: #FFFFFF;
-    font-size: 200%;
-    text-align: center;
-    padding: 5px;
-    width: 100%;
-}
-.buttainer {
-    background-color: #000000;
-    float: left;
-    height: 100px;
-    width: 100%;
-}
-.tree {
-    background-color: #DEDEDE;
-    position: relative;
-    height: 100%;
-    width: 200px;
-    float: left;
-}
-.dashboard {
-    border: 2px solid black;
-    width: 200px;
-    height: 200px;
-}
-
-.container {
-    border: 2px solid black;
-    background-color: #FF9900;
-    position: relative;
-    width: 200px;
-    height: 200px;
-}
-
-footer {
-    background-color: #000000;
-    color: #FFFFFF;
-    font-size: 200%;
-    text-align: center;
-    padding: 5px;
-    width: 100%;
-}
-*/
