--- exa/static/html/polyrefs.html
@@ -1,25 +1,0 @@
-<!-- Iron elements -->
-<link rel="import" href="../js/bower_components/iron-flex-layout/classes/iron-flex-layout.html">
-<link rel="import" href="../js/bower_components/iron-icons/iron-icons.html">
-<link rel="import" href="../js/bower_components/iron-pages/iron-pages.html">
-<link rel="import" href="../js/bower_components/iron-selector/iron-selector.html">
-
-<!-- Paper elements -->
-<link rel="import" href="../js/bower_components/paper-drawer-panel/paper-drawer-panel.html">
-<link rel="import" href="../js/bower_components/paper-icon-button/paper-icon-button.html">
-<link rel="import" href="../js/bower_components/paper-item/paper-item.html">
-<link rel="import" href="../js/bower_components/paper-material/paper-material.html">
-<link rel="import" href="../js/bower_components/paper-menu/paper-menu.html">
-<link rel="import" href="../js/bower_components/paper-scroll-header-panel/paper-scroll-header-panel.html">
-<link rel="import" href="../js/bower_components/paper-styles/paper-styles-classes.html">
-<link rel="import" href="../js/bower_components/paper-toast/paper-toast.html">
-<link rel="import" href="../js/bower_components/paper-toolbar/paper-toolbar.html">
-
-<!-- Configure your routes here -->
-<link rel="import" href="../templates/routing.html">
-
-<!-- Add your elements here -->
-<link rel="import" href="app-theme.html">
-<link rel="import" href="shared-styles.html">
-<link rel="import" href="my-greeting.html">
-<link rel="import" href="my-list.html">
