--- exa/static/html/shared-styles.html
@@ -1,20 +1,0 @@
-<!-- shared styles for all elements and index.html -->
-<dom-module id="shared-styles">
-  <template>
-    <style>
-      .page-title {
-        @apply(--paper-font-display2);
-      }
-
-      paper-menu a > *, paper-menu paper-item > *, paper-menu paper-icon-item > * {
-        pointer-events: none;
-      }
-
-      @media (max-width: 600px) {
-        .page-title {
-          font-size: 24px!important;
-        }
-      }
-    </style>
-  </template>
-</dom-module>
