--- exa/static/js/exa.dashboard.web.js
@@ -1,5 +1,0 @@
-/*"""
-Dashboard Application (Web)
-````````````````````````````
-
-*/
