+++ exa/static/nbextensions/adapters/threejs.js
@@ -1,0 +1,1 @@
+../../js/adapters/threejs.js
