+++ exa/static/nbextensions/apps/app3d.js
@@ -1,0 +1,1 @@
+../../js/apps/app3d.js
