+++ exa/static/nbextensions/tester.js
@@ -1,0 +1,1 @@
+../js/tester.js
